+++ app/src/main/java/com/example/alldayfit/MainActivity.kt
@@ -6,22 +6,21 @@ import android.view.Menu
 import com.example.alldayfit.databinding.MainActivityBinding
 
 class MainActivity : AppCompatActivity() {
-    private var mBinding: MainActivityBinding? = null
-    private val binding get() = mBinding!!
+    private lateinit var mBinding: MainActivityBinding
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         mBinding = MainActivityBinding.inflate(layoutInflater)
-        val view = binding.root
+        val view = mBinding.root
         setContentView(view)
         setCustomToolbar(R.id.main_toolbar)
     }
 
     // 툴바 메뉴 버튼을 설정- menu에 있는 item을 연결하는 부분
     fun setCustomToolbar(layout: Int) {
-        val toolbar = binding.mainToolbar
+        val toolbar = mBinding.mainToolbar
         setSupportActionBar(toolbar)
         supportActionBar?.setDisplayShowTitleEnabled(false)
-        val mTitle = binding.toolbarTitle
+        val mTitle = mBinding.toolbarTitle
         mTitle.text = "All_DayFit"
     }
