--- app/src/main/java/com/example/alldayfit/Settings/SettingMainFragment.kt
@@ -1,23 +1,0 @@
-package com.example.alldayfit.Settings
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.example.alldayfit.R
-import com.example.alldayfit.databinding.SettingMainFragmentBinding
-
-class SettingMainFragment : Fragment() {
-    private var _binding: SettingMainFragmentBinding? = null
-    private val binding get() = _binding!!
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        _binding = SettingMainFragmentBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
