--- app/src/main/java/com/example/alldayfit/Settings/SettingNoticeItem.kt
@@ -1,7 +1,0 @@
-package com.example.alldayfit.Settings
-
-data class SettingNoticeItem(
-    val notice_title: String,
-    val notice_time: Int,
-    val notice_info: String
-)
