+++ app/src/main/java/com/example/alldayfit/calendar/CalendarActivity.kt
@@ -1,11 +1,8 @@
-package com.example.alldayfit
+package com.example.alldayfit.calendar
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
 import androidx.appcompat.app.AppCompatActivity
+import com.example.alldayfit.R
 import com.example.alldayfit.databinding.CalendarActivityBinding
