+++ app/src/main/java/com/example/alldayfit/calendar/CalendarFragment.kt
@@ -1,0 +1,27 @@
+package com.example.alldayfit.calendar
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.example.alldayfit.databinding.CalendarFragmentBinding
+
+class CalendarFragment : Fragment() {
+
+    private var _binding: CalendarFragmentBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = CalendarFragmentBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
