+++ app/src/main/java/com/example/alldayfit/community/CommunityHomeFragment.kt
@@ -1,0 +1,29 @@
+package com.example.alldayfit.community
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.example.alldayfit.databinding.CalendarFragmentBinding
+import com.example.alldayfit.databinding.CommunityHomeFragmentBinding
+
+class CommunityHomeFragment : Fragment() {
+    private var _binding: CommunityHomeFragmentBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = CommunityHomeFragmentBinding.inflate(inflater, container, false)
+        val view = binding.root
+        return view
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
