+++ app/src/main/java/com/example/alldayfit/dietrecord/DietRecordFragment.kt
@@ -1,4 +1,4 @@
-package com.example.alldayfit
+package com.example.alldayfit.dietrecord
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
