+++ app/src/main/java/com/example/alldayfit/main/MainFragment.kt
@@ -1,22 +1,28 @@
-package com.example.alldayfit
+package com.example.alldayfit.main
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.alldayfit.databinding.CalendarFragmentBinding
+import com.example.alldayfit.databinding.FragmentMainBinding
+import com.example.alldayfit.databinding.MainActivityBinding
 
-class CalendarFragment : Fragment() {
 
-    private var _binding: CalendarFragmentBinding? = null
+class MainFragment : Fragment() {
+    private var _binding: FragmentMainBinding? = null
     private val binding get() = _binding!!
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+    }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = CalendarFragmentBinding.inflate(inflater, container, false)
+        _binding = FragmentMainBinding.inflate(inflater, container, false)
         return binding.root
     }
 
@@ -24,4 +30,5 @@ class CalendarFragment : Fragment() {
         super.onDestroyView()
         _binding = null
     }
+
 }
