+++ app/src/main/java/com/example/alldayfit/settings/adapter/SettingNoticeAdapter.kt
@@ -1,18 +1,21 @@
-package com.example.alldayfit.Settings
+package com.example.alldayfit.settings.adapter
 
 import android.content.Context
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
+import com.example.alldayfit.settings.ui.SettingNoticeItem
 import com.example.alldayfit.databinding.SettingNoticeItemBinding
 
-class SettingNoticeAdapter(private val context: Context) : RecyclerView.Adapter<SettingNoticeAdapter.Holder>() {
+class SettingNoticeAdapter(private val context: Context) :
+    RecyclerView.Adapter<SettingNoticeAdapter.Holder>() {
 
-    private val items : MutableList<SettingNoticeItem> = mutableListOf()
+    private val items: MutableList<SettingNoticeItem> = mutableListOf()
     var itemsNotice = ArrayList<SettingNoticeItem>()
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): Holder {
-        val binding = SettingNoticeItemBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+        val binding =
+            SettingNoticeItemBinding.inflate(LayoutInflater.from(parent.context), parent, false)
         return Holder(binding)
     }
 
@@ -24,7 +27,8 @@ class SettingNoticeAdapter(private val context: Context) : RecyclerView.Adapter<
         return items.size
     }
 
-    inner class Holder(val binding: SettingNoticeItemBinding) : RecyclerView.ViewHolder(binding.root) {
+    inner class Holder(val binding: SettingNoticeItemBinding) :
+        RecyclerView.ViewHolder(binding.root) {
 
     }
