+++ app/src/main/java/com/example/alldayfit/settings/ui/SettingAppNotifiFragment.kt
@@ -1,23 +1,26 @@
-package com.example.alldayfit.Settings
+package com.example.alldayfit.settings.ui
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.alldayfit.R
 import com.example.alldayfit.databinding.SettingNotifiFragmentBinding
 
 class SettingAppNotifiFragment : Fragment() {
     private var _binding: SettingNotifiFragmentBinding? = null
     private val binding get() = _binding!!
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        _binding = SettingNotifiFragmentBinding.inflate(inflater, container,false)
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = SettingNotifiFragmentBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onDestroyView() {
         super.onDestroyView()
-        _binding
+        _binding = null
     }
 }
