+++ app/src/main/java/com/example/alldayfit/settings/ui/SettingModUserFragment.kt
@@ -1,17 +1,20 @@
-package com.example.alldayfit.Settings
+package com.example.alldayfit.settings.ui
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.alldayfit.R
 import com.example.alldayfit.databinding.SettingModUserFragmentBinding
 
 class SettingModUserFragment : Fragment() {
     private var _binding: SettingModUserFragmentBinding? = null
     private val binding get() = _binding!!
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
         _binding = SettingModUserFragmentBinding.inflate(inflater, container, false)
         return binding.root
     }
