+++ app/src/main/java/com/example/alldayfit/settings/ui/SettingNoticeItem.kt
@@ -1,0 +1,7 @@
+package com.example.alldayfit.settings.ui
+
+data class SettingNoticeItem(
+    val noticeTitle: String,
+    val noticeTime: Int,
+    val noticeInfo: String
+)
