+++ app/src/main/res/layout/calendar_activity.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/light_blue"
-    tools:context=".CalendarActivity">
+    tools:context=".calendar.CalendarActivity">
 
     <TextView
         android:id="@+id/textView"
