+++ app/src/main/res/layout/calendar_day_legend_container.xml
@@ -1,8 +1,8 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:weightSum="7"
-    android:orientation="horizontal">
+    android:orientation="horizontal"
+    android:weightSum="7">
 
     <include layout="@layout/calendar_day_legend_text" />
