+++ app/src/main/res/layout/calendar_header.xml
@@ -39,7 +39,6 @@
         app:layout_constraintTop_toTopOf="parent" />
 
 
-
     <View
         android:layout_width="0dp"
         android:layout_height="0.75dp"
