+++ app/src/main/res/layout/diet_record_fragment.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="15dp"
-    tools:context=".DietRecordFragment">
+    tools:context=".dietrecord.DietRecordFragment">
 
     <TextView
         android:id="@+id/date_txt"
