+++ app/src/main/res/layout/exercise_status_fragment.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/blue"
-    tools:context=".ExerciseStatusFragment">
+    tools:context=".exercisestatus.ExerciseStatusFragment">
 
     <TextView
         android:id="@+id/date_txt"
