+++ app/src/main/res/layout/exercise_status_user_item.xml
@@ -1,30 +1,30 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="100dp"
     android:layout_height="100dp"
-    android:background="@drawable/exercise_status_user_item_shape"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
+    android:background="@drawable/exercise_status_user_item_shape">
 
     <TextView
         android:id="@+id/status_type_txt"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
         android:layout_marginTop="15dp"
-        android:text="@string/weight"/>
+        android:text="@string/weight"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
         android:id="@+id/status_metrics_txt"
         android:layout_width="70dp"
         android:layout_height="50dp"
-        android:textColor="@color/navy"
-        android:gravity="center"
         android:background="@color/white"
+        android:gravity="center"
         android:text="70kg"
-        app:layout_constraintStart_toStartOf="parent"
+        android:textColor="@color/navy"
+        app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"/>
+        app:layout_constraintStart_toStartOf="parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
