+++ app/src/main/res/layout/fragment_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/blue"
-    tools:context=".MainFragment">
+    tools:context=".main.MainFragment">
 
 
     <TextView
