+++ app/src/main/res/layout/setting_mod_user_fragment.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:paddingTop="15dp"
     android:padding="10dp"
-    tools:context=".Settings.SettingModUserFragment">
+    tools:context=".settings.ui.SettingModUserFragment">
 
     <TextView
         android:id="@+id/textView_1"
