+++ app/src/main/res/layout/setting_notice_fragment.xml
@@ -7,7 +7,7 @@
     android:orientation="vertical"
     android:paddingTop="15dp"
     android:padding="10dp"
-    tools:context=".Settings.SettingNoticeFragment">
+    tools:context=".settings.ui.SettingNoticeFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recycler_view"
