+++ app/src/main/res/layout/setting_notifi_fragment.xml
@@ -6,7 +6,7 @@
     android:orientation="vertical"
     android:paddingHorizontal="10dp"
     android:paddingTop="15dp"
-    tools:context=".Settings.SettingMainFragment">
+    tools:context=".settings.ui.SettingMainFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
